+++ dialect.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"strings"
-
-	"github.com/xo/dburl"
+	"github.com/hundredwatt/teleport/schema"
 )
 
 type Dialect struct {
@@ -13,57 +11,61 @@ type Dialect struct {
 	CreateStagingTableQuery string
 	FullLoadQuery           string
 	ModifiedOnlyLoadQuery   string
+	SetSchemaQuery          string
 }
 
 var (
-	mysql    = Dialect{"mysql", "MySQL", "", "", "", ""}
+	mysql    = Dialect{"mysql", "MySQL", "", "", "", "", ""}
 	postgres = Dialect{"postgres", "PostgreSQL", "psql",
 		"CREATE TABLE %[2]s AS TABLE %[1]s WITH NO DATA",
 		`
-			ALTER TABLE %[1]s RENAME TO archive_%[1]s;
-			ALTER TABLE %[2]s RENAME TO %[1]s;
-			DROP TABLE archive_%[1]s;
+			DELETE FROM %[1]s;
+			INSERT INTO %[1]s SELECT * FROM %[2]s
 		`,
 		`
 			DELETE FROM %[1]s WHERE %[3]s IN (SELECT %[3]s FROM %[2]s);
 			INSERT INTO %[1]s SELECT * FROM %[2]s;
 			DROP TABLE %[2]s
-		`}
+		`,
+		"SET search_path TO %s"}
 	redshift = Dialect{"redshift", "AWS RedShift", "psql",
 		"CREATE TEMPORARY TABLE %[2]s (LIKE %[1]s)",
 		`
-			BEGIN TRANSACTION;
 			DELETE FROM %[1]s;
-			INSERT INTO %[1]s SELECT * FROM %[2]s;
-			END TRANSACTION;
+			INSERT INTO %[1]s SELECT * FROM %[2]s
 		`,
 		`
 			DELETE FROM %[1]s USING %[2]s WHERE %[1]s.%[3]s = %[2]s.%[3]s;
-			INSERT INTO %[1]s SELECT * FROM %[2]s;
-		`}
+			INSERT INTO %[1]s SELECT * FROM %[2]s
+		`,
+		postgres.SetSchemaQuery}
+	snowflake = Dialect{"snowflake", "Snowflake", "psql",
+		"CREATE TEMPORARY TABLE %[2]s LIKE %[1]s",
+		`
+			TRUNCATE TABLE %[1]s;
+			INSERT INTO %[1]s SELECT * FROM %[2]s
+		`,
+		`
+			MERGE INTO %[1]s USING %[2]s ON %[1]s.%[3]s = %[2]s.%[3]s;
+		`,
+		"USE SCHEMA %s"}
 	sqlite = Dialect{"sqlite", "SQLite3", "",
-		"CREATE TABLE %[2]s AS SELECT * FROM %[1]s LIMIT 0", postgres.FullLoadQuery, postgres.ModifiedOnlyLoadQuery}
+		"CREATE TABLE %[2]s AS SELECT * FROM %[1]s LIMIT 0", postgres.FullLoadQuery, postgres.ModifiedOnlyLoadQuery, ""}
 )
 
-func GetDialect(d Database) Dialect {
-	if strings.HasPrefix(d.URL, "redshift://") {
+func GetDialect(db *schema.Database) Dialect {
+	switch db.Driver {
+	case "postgres":
+		return postgres
+	case "sqlite3":
+		return sqlite
+	case "mysql":
+		return mysql
+	case "redshift":
 		return redshift
-	} else if strings.HasPrefix(d.URL, "rs://") {
-		return redshift
+	case "snowflake":
+		return snowflake
+	default:
+		return mysql
 	}
-
-	if u, err := dburl.Parse(d.URL); err == nil {
-		switch u.Driver {
-		case "postgres":
-			return postgres
-		case "sqlite3":
-			return sqlite
-		case "mysql":
-			return mysql
-		default:
-			return mysql
-		}
-	}
-
-	return mysql
 }